scientific notation progress

+++ srcs/ft_pow.c
@@ -6,13 +6,13 @@
 /*   By: rreedy <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/18 17:17:04 by rreedy            #+#    #+#             */
-/*   Updated: 2018/12/18 17:26:08 by rreedy           ###   ########.fr       */
+/*   Updated: 2019/01/02 19:43:56 by rreedy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_pow(int n, int exponent)
+size_t	ft_pow(int n, int exponent)
 {
 	long	power;
